+++ CHANGELOG.md
@@ -34,6 +34,7 @@ With this method, the version is YEAR.MONTH.RELEASE. To increment this version, 
 - Interactive & Template: Make context optional.
 - Interactive: Allow context to be added inline with `/context <context>`.
 - Interactive: A blank context, unsets the context.
+- Ollama: Test both generate and chat API.
 
 ### Changed